Add files via upload
This Jupiter notebook contains Python implementation of Linked List. I first create a singly linked list and define operations such as insertion at the beginning of the lis, insertion at the end of the list etc. I create a doubly linked list and create functions for the same operations. 
Each operation is followed with one demonstration.